message/chat counter & number

+++ app/controllers/api/v1/messages_controller.rb
@@ -14,32 +14,43 @@ class Api::V1::MessagesController < ApplicationController
             render(json: {msg: "application not found"}, status: :not_found)
             return
         end
-        # @chat = @application.chats.find_by(number: params[:number])
-        @chat = @application.chats.find(params[:number])
+
+        @chat = @application.chats.find_by(number: params[:chat_number])
+
         if @chat == nil 
             render(json: {msg: "chat not found"}, status: :not_found)
             return
         end
 
-        @message = @chat.messages.create(content: params[:content], number: 4)
-
-        # update messages_count column
-        messages_count = @chat.messages.count
-        @chat.update(messages_count: messages_count)
+        @message = @chat.messages.create(content: params[:content])
 
         if @message
             render(json: {"number": @message.number}, status: :created)
+            return
         else
             render(json: {}, status: :internal_server_error)
+            return
         end
     end
 
     def delete
         @application = Application.find_by(token: params[:token])
-        # @chat = @application.chats.find_by(number: params[:chat_number])
-        @chat = @application.chats.find(params[:chat_number])
-        # @message = @chat.messages.find_by(number: params[:message_number])
-        @message = @chat.messages.find(params[:message_number])
+        if @application == nil
+            render(json: {"error": "application not found"}, status: :not_found)
+            return
+        end
+
+        @chat = @application.chats.find_by(number: params[:chat_number])
+        if @chat == nil
+            render(json: {"error":"chat not found"}, status: :not_found)
+            return
+        end
+
+        @message = @chat.messages.find_by(number: params[:message_number])
+        if @message == nil
+            render(json: {"error": "message not found" }, status: :not_found)
+            return
+        end
 
         if @message.destroy
             render(json: {}, status: :ok)
